% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aesInput.r
\name{aesInput}
\alias{aesInput}
\title{Creates an aesInput control on a Shiny UI.}
\usage{
aesInput(
  inputId,
  label,
  data,
  allow = c("factor", "numeric"),
  picker = "basic",
  vs = FALSE
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{data}{The data.frame to use for populating the selectize dropdown.}

\item{allow}{What types of column from \code{data} to use.
Default: \code{c("factor", "numeric")}.}

\item{picker}{\code{"basic"} (default), \code{"multi"}, \code{"color"},
\code{"shape"}, or \code{"pattern"}.}

\item{vs}{If \code{TRUE}, then instead of showing the values in the column,
combinations of values will be displayed instead.}
}
\value{
A list of aesInput values.
}
\description{
Creates an aesInput control on a Shiny UI.
}
\seealso{
Other aesInput: 
\code{\link{ai.aes}()},
\code{\link{ai.column}()},
\code{\link{ai.flatten}()},
\code{\link{ai.scales}()},
\code{\link{ai.subset}()},
\code{\link{ai.values}()}
}
\concept{aesInput}
