% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.r
\name{ai.aes}
\alias{ai.aes}
\title{Generate aesthetic mappings based on one or more aesInputs.}
\usage{
ai.aes(..., exec = TRUE)
}
\arguments{
\item{...}{aesInput values, arbitrarily nested in lists. Other 
\link[ggplot2]{aes_string} arguments can go here as well and will
override those from aesInputs.}

\item{exec}{When \code{TRUE} (the default) an aes object will be returned,
otherwise a list of aes key/value pairs will be returned.}
}
\value{
See \code{exec}, above.
}
\description{
Generate aesthetic mappings based on one or more aesInputs.
}
\seealso{
Other aesInput: 
\code{\link{aesInput}()},
\code{\link{ai.flatten}()},
\code{\link{ai.scales}()},
\code{\link{ai.subset}()}
}
\concept{aesInput}
