% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.r
\name{ai.column}
\alias{ai.column}
\title{Retrieve the selected column name.}
\usage{
ai.column(ai)
}
\arguments{
\item{ai}{A single aesInput object}
}
\value{
The name of the selected column.
}
\description{
Retrieve the selected column name.
}
\seealso{
Other aesInput: 
\code{\link{aesInput}()},
\code{\link{ai.aes}()},
\code{\link{ai.flatten}()},
\code{\link{ai.scales}()},
\code{\link{ai.subset}()},
\code{\link{ai.values}()}
}
\concept{aesInput}
