% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.r
\name{ai.flatten}
\alias{ai.flatten}
\title{Turns nested lists of aesInputs into a single list.}
\usage{
ai.flatten(...)
}
\arguments{
\item{...}{aesInput values, arbitrarily nested in lists.}
}
\value{
A list of aesInput values.
}
\description{
Turns nested lists of aesInputs into a single list.
}
\seealso{
Other aesInput: 
\code{\link{aesInput}()},
\code{\link{ai.aes}()},
\code{\link{ai.scales}()},
\code{\link{ai.subset}()}
}
\concept{aesInput}
