% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.r
\name{ai.scales}
\alias{ai.scales}
\title{Assign colors/shapes/patterns to data values using ggplot2 or ggpattern 
scales.}
\usage{
ai.scales(..., exec = TRUE)
}
\arguments{
\item{...}{aesInput values, arbitrarily nested in lists. Other 
\link[ggplot2]{aes_string} arguments can go here as well and will
override those from aesInputs.}

\item{exec}{If exec is TRUE (the default), returns a ggproto object, a list
describing what functions should be run with what parameters is
returned.}
}
\value{
A list of ggproto objects, or a descriptive list thereof if 
        \code{exec=FALSE}.
}
\description{
Assign colors/shapes/patterns to data values using ggplot2 or ggpattern 
scales.
}
\seealso{
Other aesInput: 
\code{\link{aesInput}()},
\code{\link{ai.aes}()},
\code{\link{ai.column}()},
\code{\link{ai.flatten}()},
\code{\link{ai.subset}()},
\code{\link{ai.values}()}
}
\concept{aesInput}
