% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.r
\name{ai.subset}
\alias{ai.subset}
\title{Subset a data.frame based on aesInput(s) values.}
\usage{
ai.subset(data, ..., exec = TRUE)
}
\arguments{
\item{data}{A data.frame/tibble/etc object.}

\item{...}{aesInput values, arbitrarily nested in lists.}

\item{exec}{When \code{TRUE} (the default) returns a subsetted version of 
\code{data}. If \code{FALSE}, a logical expression is returned 
instead.}
}
\value{
Subsetted \code{data} object or a logical expression. See 
        \code{exec}, above.
}
\description{
Subset a data.frame based on aesInput(s) values.
}
\seealso{
Other aesInput: 
\code{\link{aesInput}()},
\code{\link{ai.aes}()},
\code{\link{ai.column}()},
\code{\link{ai.flatten}()},
\code{\link{ai.scales}()},
\code{\link{ai.values}()}
}
\concept{aesInput}
