% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.r
\name{ai.values}
\alias{ai.values}
\title{Retrieve the selected column's values.}
\usage{
ai.values(ai)
}
\arguments{
\item{ai}{A single aesInput object.}
}
\value{
A named character vector. Names are the values from the column; 
        values are the additional attribute (color, shape, pattern). When
        \code{picker = "basic"} or \code{picker = "multi"} keys and values 
        will be identical.
}
\description{
Retrieve the selected column's values.
}
\seealso{
Other aesInput: 
\code{\link{aesInput}()},
\code{\link{ai.aes}()},
\code{\link{ai.column}()},
\code{\link{ai.flatten}()},
\code{\link{ai.scales}()},
\code{\link{ai.subset}()}
}
\concept{aesInput}
