% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai.r
\name{ai.subset}
\alias{ai.subset}
\title{Subset a data.frame based on aesInput(s) values.}
\usage{
ai.subset(data, ...)
}
\arguments{
\item{data}{A data.frame/tibble/etc object.}

\item{...}{aesInput values, arbitrarily nested in lists.}
}
\value{
Subsetted \code{data} object.
}
\description{
Subset a data.frame based on aesInput(s) values.
}
\seealso{
Other aesInput: 
\code{\link{ai.aes}()},
\code{\link{ai.flatten}()},
\code{\link{ai.scales}()}
}
\concept{aesInput}
